+++ example/lib/pages/room.dart
@@ -54,9 +54,15 @@ class _RoomPageState extends State<RoomPage> {
   }
 
   void _setUpListeners() => _listener
-    ..on<RoomDisconnectedEvent>((_) async {
+    ..on<RoomDisconnectedEvent>((event) async {
+      if (event.reason != null) {
+        print('Room disconnected: reason => ${event.reason}');
+      }
       WidgetsBindingCompatible.instance
           ?.addPostFrameCallback((timeStamp) => Navigator.pop(context));
+    })
+    ..on<RoomRecordingStatusChanged>((event) {
+      context.showRecordingStatusChangedDialog(event.activeRecording);
     })
     ..on<LocalTrackPublishedEvent>((_) => _sortParticipants())
     ..on<LocalTrackUnpublishedEvent>((_) => _sortParticipants())