+++ lib/src/core/engine.dart
@@ -19,6 +19,7 @@ import '../proto/livekit_models.pb.dart' as lk_models;
 import '../proto/livekit_rtc.pb.dart' as lk_rtc;
 import '../support/disposable.dart';
 import '../support/websocket.dart';
+import '../types/internal.dart';
 import '../types/other.dart';
 import '../types/video_dimensions.dart';
 import '../utils.dart';
@@ -237,7 +238,7 @@ class Engine extends Disposable with EventsEmittable<EngineEvent> {
       if (error is NegotiationError) {
         fullReconnect = true;
       }
-      await handleDisconnect(DisconnectReason.negotiationFailed);
+      await handleDisconnect(ClientDisconnectReason.negotiationFailed);
     }
   }
 
@@ -295,7 +296,7 @@ class Engine extends Disposable with EventsEmittable<EngineEvent> {
   }
 
   Future<void> _configurePeerConnections(
-      {required lk_models.ClientConfigSetting forceRelay,
+      {required lk_models.ClientConfigSetting serverResponseForceRelay,
       required List<RTCIceServer> serverProvidedIceServers}) async {
     if (publisher != null || subscriber != null) {
       logger.warning('Already configured');
@@ -313,23 +314,11 @@ class Engine extends Disposable with EventsEmittable<EngineEvent> {
           .copyWith(iceServers: serverProvidedIceServers);
     }
 
-    // set forceRelay
-    if (rtcConfiguration.iceTransportPolicy == null) {
-      switch (forceRelay) {
-        case lk_models.ClientConfigSetting.ENABLED:
-          rtcConfiguration = rtcConfiguration.copyWith(
-            iceTransportPolicy: RTCIceTransportPolicy.relay,
-          );
-          break;
-        case lk_models.ClientConfigSetting.DISABLED:
-          rtcConfiguration = rtcConfiguration.copyWith(
-            iceTransportPolicy: RTCIceTransportPolicy.all,
-          );
-          break;
-        case lk_models.ClientConfigSetting.UNSET:
-          // do nothing
-          break;
-      }
+    // set forceRelay if server response is enabled
+    if (serverResponseForceRelay == lk_models.ClientConfigSetting.ENABLED) {
+      rtcConfiguration = rtcConfiguration.copyWith(
+        iceTransportPolicy: RTCIceTransportPolicy.relay,
+      );
     }
 
     if (kIsWeb && roomOptions.e2eeOptions != null) {
@@ -392,9 +381,9 @@ class Engine extends Disposable with EventsEmittable<EngineEvent> {
 
     events.on<EnginePeerStateUpdatedEvent>((event) {
       if (event.state.isDisconnectedOrFailed()) {
-        handleDisconnect(DisconnectReason.reconnect);
+        handleDisconnect(ClientDisconnectReason.reconnect);
       } else if (event.state.isClosed()) {
-        handleDisconnect(DisconnectReason.peerConnectionClosed);
+        handleDisconnect(ClientDisconnectReason.peerConnectionClosed);
       }
     });
 
@@ -558,7 +547,7 @@ class Engine extends Disposable with EventsEmittable<EngineEvent> {
     }
   }
 
-  Future<void> handleDisconnect(DisconnectReason reason) async {
+  Future<void> handleDisconnect(ClientDisconnectReason reason) async {
     logger
         .info('onDisconnected state:${_connectionState} reason:${reason.name}');
 
@@ -566,9 +555,9 @@ class Engine extends Disposable with EventsEmittable<EngineEvent> {
       fullReconnect = _clientConfiguration?.resumeConnection ==
               lk_models.ClientConfigSetting.DISABLED ||
           [
-            DisconnectReason.leaveReconnect,
-            DisconnectReason.negotiationFailed,
-            DisconnectReason.peerConnectionClosed
+            ClientDisconnectReason.leaveReconnect,
+            ClientDisconnectReason.negotiationFailed,
+            ClientDisconnectReason.peerConnectionClosed
           ].contains(reason);
     }
 
@@ -744,7 +733,8 @@ class Engine extends Disposable with EventsEmittable<EngineEvent> {
           'forceRelay: $event.response.clientConfiguration.forceRelay');
 
       await _configurePeerConnections(
-          forceRelay: event.response.clientConfiguration.forceRelay,
+          serverResponseForceRelay:
+              event.response.clientConfiguration.forceRelay,
           serverProvidedIceServers: _serverProvidedIceServers);
 
       if (!_subscriberPrimary) {
@@ -754,7 +744,7 @@ class Engine extends Disposable with EventsEmittable<EngineEvent> {
     })
     ..on<SignalConnectionStateUpdatedEvent>((event) async {
       if (event.newState == ConnectionState.disconnected) {
-        await handleDisconnect(DisconnectReason.signal);
+        await handleDisconnect(ClientDisconnectReason.signal);
       }
     })
     ..on<SignalOfferEvent>((event) async {
@@ -810,14 +800,15 @@ class Engine extends Disposable with EventsEmittable<EngineEvent> {
         // reconnect immediately instead of waiting for next attempt
         _connectionState = ConnectionState.reconnecting;
         _updateConnectionState(ConnectionState.reconnecting);
-        await handleDisconnect(DisconnectReason.leaveReconnect);
+        await handleDisconnect(ClientDisconnectReason.leaveReconnect);
       } else {
         if (_connectionState == ConnectionState.reconnecting) {
           logger.warning(
               '[Signal] Received Leave while engine is reconnecting, ignoring...');
           return;
         }
-        _updateConnectionState(ConnectionState.disconnected);
+        _updateConnectionState(ConnectionState.disconnected,
+            reason: event.reason.toSDKType());
         await cleanUp();
       }
     });
@@ -833,7 +824,8 @@ extension EnginePrivateMethods on Engine {
       _publisherDataChannel(reliability)?.state ??
       rtc.RTCDataChannelState.RTCDataChannelClosed;
 
-  void _updateConnectionState(ConnectionState newValue) {
+  void _updateConnectionState(ConnectionState newValue,
+      {DisconnectReason? reason}) {
     if (_connectionState == newValue) return;
 
     logger.fine('Engine ConnectionState '
@@ -850,6 +842,7 @@ extension EnginePrivateMethods on Engine {
       oldState: oldState,
       didReconnect: didReconnect,
       fullReconnect: fullReconnect,
+      disconnectReason: reason,
     ));
   }
 }