+++ lib/src/core/room.dart
@@ -2,6 +2,7 @@ import 'dart:async';
 
 import 'package:collection/collection.dart';
 import 'package:flutter/foundation.dart';
+import 'package:livekit_client/src/hardware/hardware.dart';
 import 'package:livekit_client/src/support/app_state.dart';
 import 'package:meta/meta.dart';
 
@@ -73,6 +74,8 @@ class Room extends DisposableChangeNotifier with EventsEmittable<RoomEvent> {
   String? _serverRegion;
 
   E2EEManager? _e2eeManager;
+  bool get isRecording => _isRecording;
+  bool _isRecording = false;
 
   /// a list of participants that are actively speaking, including local participant.
   UnmodifiableListView<Participant> get activeSpeakers =>
@@ -164,6 +167,12 @@ class Room extends DisposableChangeNotifier with EventsEmittable<RoomEvent> {
       _metadata = event.response.room.metadata;
       _serverVersion = event.response.serverVersion;
       _serverRegion = event.response.serverRegion;
+
+      if (_isRecording != event.response.room.activeRecording) {
+        _isRecording = event.response.room.activeRecording;
+        emitWhenConnected(
+            RoomRecordingStatusChanged(activeRecording: _isRecording));
+      }
 
       logger.fine('[Engine] Received JoinResponse, '
           'serverVersion: ${event.response.serverVersion}');
@@ -273,6 +282,11 @@ class Room extends DisposableChangeNotifier with EventsEmittable<RoomEvent> {
       _metadata = event.room.metadata;
       emitWhenConnected(
           RoomMetadataChangedEvent(metadata: event.room.metadata));
+      if (_isRecording != event.room.activeRecording) {
+        _isRecording = event.room.activeRecording;
+        emitWhenConnected(
+            RoomRecordingStatusChanged(activeRecording: _isRecording));
+      }
     })
     ..on<SignalConnectionStateUpdatedEvent>((event) {
       // during reconnection, need to send sync state upon signal connection.
@@ -325,7 +339,7 @@ class Room extends DisposableChangeNotifier with EventsEmittable<RoomEvent> {
       } else if (event.newState == ConnectionState.disconnected) {
         if (!event.fullReconnect) {
           await _cleanUp();
-          events.emit(const RoomDisconnectedEvent());
+          events.emit(RoomDisconnectedEvent(reason: event.disconnectReason));
         }
       }
       // always notify ChangeNotifier
@@ -355,6 +369,7 @@ class Room extends DisposableChangeNotifier with EventsEmittable<RoomEvent> {
           event.stream,
           trackSid,
           receiver: event.receiver,
+          audioOutputOptions: roomOptions.defaultAudioOutputOptions,
         );
       } on TrackSubscriptionExceptionEvent catch (event) {
         logger.severe('addSubscribedMediaTrack() throwed ${event}');
@@ -658,5 +673,103 @@ extension RoomDebugMethods on Room {
         migration: migration,
         serverLeave: serverLeave,
         switchCandidate: switchCandidate);
+  }
+}
+
+/// Room extension methods for managing audio, video.
+extension RoomHardwareManagementMethods on Room {
+  /// Get current audio output device.
+  String? get selectedAudioOutputDeviceId =>
+      roomOptions.defaultAudioOutputOptions.deviceId ??
+      Hardware.instance.selectedAudioOutput?.deviceId;
+
+  /// Get current audio input device.
+  String? get selectedAudioInputDeviceId =>
+      roomOptions.defaultAudioCaptureOptions.deviceId ??
+      Hardware.instance.selectedAudioInput?.deviceId;
+
+  /// Get current video input device.
+  String? get selectedVideoInputDeviceId =>
+      roomOptions.defaultCameraCaptureOptions.deviceId ??
+      Hardware.instance.selectedVideoInput?.deviceId;
+
+  /// Get mobile device's speaker status.
+  bool? get speakerOn => roomOptions.defaultAudioOutputOptions.speakerOn;
+
+  /// Set audio output device.
+  Future<void> setAudioOutputDevice(MediaDevice device) async {
+    if (lkPlatformIs(PlatformType.web)) {
+      participants.forEach((_, participant) {
+        for (var audioTrack in participant.audioTracks) {
+          audioTrack.track?.setSinkId(device.deviceId);
+        }
+      });
+      Hardware.instance.selectedAudioOutput = device;
+    } else {
+      await Hardware.instance.selectAudioOutput(device);
+    }
+    engine.roomOptions = engine.roomOptions.copyWith(
+      defaultAudioOutputOptions: roomOptions.defaultAudioOutputOptions.copyWith(
+        deviceId: device.deviceId,
+      ),
+    );
+  }
+
+  /// Set audio input device.
+  Future<void> setAudioInputDevice(MediaDevice device) async {
+    if (lkPlatformIs(PlatformType.web) && localParticipant != null) {
+      for (var audioTrack in localParticipant!.audioTracks) {
+        await audioTrack.track?.setDeviceId(device.deviceId);
+      }
+      Hardware.instance.selectedAudioInput = device;
+    } else {
+      await Hardware.instance.selectAudioInput(device);
+    }
+    engine.roomOptions = engine.roomOptions.copyWith(
+      defaultAudioCaptureOptions:
+          roomOptions.defaultAudioCaptureOptions.copyWith(
+        deviceId: device.deviceId,
+      ),
+    );
+  }
+
+  /// Set video input device.
+  Future<void> setVideoInputDevice(MediaDevice device) async {
+    final track = localParticipant?.videoTracks.firstOrNull?.track;
+    if (track == null) return;
+    if (selectedVideoInputDeviceId != device.deviceId) {
+      await track.switchCamera(device.deviceId);
+      Hardware.instance.selectedVideoInput = device;
+    }
+    engine.roomOptions = engine.roomOptions.copyWith(
+      defaultCameraCaptureOptions:
+          roomOptions.defaultCameraCaptureOptions.copyWith(
+        deviceId: device.deviceId,
+      ),
+    );
+  }
+
+  Future<void> setSpeakerOn(bool speakerOn) async {
+    if (lkPlatformIs(PlatformType.iOS) || lkPlatformIs(PlatformType.android)) {
+      await Hardware.instance.setSpeakerphoneOn(speakerOn);
+      engine.roomOptions = engine.roomOptions.copyWith(
+        defaultAudioOutputOptions:
+            roomOptions.defaultAudioOutputOptions.copyWith(
+          speakerOn: speakerOn,
+        ),
+      );
+    }
+  }
+
+  /// Apply audio output device settings.
+  @internal
+  Future<void> applyAudioSpeakerSettings() async {
+    if (roomOptions.defaultAudioOutputOptions.speakerOn != null) {
+      if (lkPlatformIs(PlatformType.iOS) ||
+          lkPlatformIs(PlatformType.android)) {
+        await Hardware.instance.setSpeakerphoneOn(
+            roomOptions.defaultAudioOutputOptions.speakerOn!);
+      }
+    }
   }
 }