+++ lib/src/events.dart
@@ -65,10 +65,13 @@ class RoomRestartedEvent with RoomEvent {
 /// Disconnected from the room
 /// Emitted by [Room].
 class RoomDisconnectedEvent with RoomEvent {
-  const RoomDisconnectedEvent();
+  DisconnectReason? reason;
+  RoomDisconnectedEvent({
+    this.reason,
+  });
 
   @override
-  String toString() => '${runtimeType}()';
+  String toString() => '${runtimeType}($reason)';
 }
 
 /// Room metadata has changed.
@@ -82,6 +85,19 @@ class RoomMetadataChangedEvent with RoomEvent {
 
   @override
   String toString() => '${runtimeType}()';
+}
+
+/// Room recording status has changed.
+/// Emitted by [Room].
+class RoomRecordingStatusChanged with RoomEvent {
+  final bool activeRecording;
+
+  const RoomRecordingStatusChanged({
+    required this.activeRecording,
+  });
+
+  @override
+  String toString() => '${runtimeType}(activeRecording = $activeRecording)';
 }
 
 /// When a new [RemoteParticipant] joins *after* the current participant has connected